# config/swagger.yml
swagger: "2.0"
info:
  description: This is the swagger file that goes with our server code
  version: "1.0.0"
  title: Swagger REST Article
consumes:
  - application/json
produces:
  - application/json

basePath: /api

# Paths supported by the server application
paths:
  /intersite-vertical:
    get:
      operationId: vertical_read_all_service
      tags:
        - services
      summary: the inter-site Service mapping structure GET
      description: Read the list of inter-site services
      x-swagger-router-controller: controllers.intersite
      responses:
        200:
          description: Successful read inter-site service list operation
          schema:
            type: array
            items:
              properties:
                service_id:
                  type: string
                  description: ID of the inter-site service
                service_type:
                  type: string
                  description: Inter-site service type (L2 or L3)
                service_name:
                  type: string
                  description: Name of the inter-site service
                service_params:
                  type: string
                  description: Additional params for the inter-site service
                service_resources:
                  type: array
                  description: Resources to federate
                  items: 
                    properties:
                      service_id:
                        type: integer
                      resource_region:    
                        type: string
                      resource_uuid:    
                        type: string  
                service_interconnections:
                  type: array
                  description: Locally created interconnections
                  items:
                    properties:
                      interconnexion_uuid:
                        type: string
    post:
      operationId: vertical_create_service
      tags:
        - services
      summary: Creates an inter-site Service POST
      description: Create an inter-site service
      x-swagger-router-controller: controllers.intersite
      parameters:
        - name: service
          in: body
          description: data for inter-service creation
          required: True
          schema:
            type: object
            properties:
              type:
                type: string
                description: Type of inter-site service
              name:
                type: string
                description: Name of the inter-site service
              resources:
                type: array
                items:
                  type: string 
      responses:
        201:
          description: Successful create inter-site service operation
          schema:
            properties:
              id:
                type: string
                description: Id of the person
              type:
                type: string
                description: Inter-site service type
              name:
                type: string
                description: Name of the inter-site service
              resources:
                type: array
                description: Resources to federate
                items: 
                  type: string  
              interconnections:
                type: array
                description: Locally created interconnections
                items: 
                  type: string
    
  /intersite-vertical/{global_id}:
    delete:
      operationId: vertical_delete_service
      tags:
        - services
      summary: Deletes an inter-site Service DELETE
      description: Deletes an inter-site service
      x-swagger-router-controller: controllers.intersite
      parameters:
        - name: global_id
          in: path
          type: string
          required: True
          description: Id of the service to delete
      responses:
        200:
          description: Successfully deleted a service from service list

    get:
      operationId: vertical_read_one_service
      tags:
        - services
      summary: Read one inter-site Service GET
      description: Read one inter-site service
      x-swagger-router-controller: controllers.intersite
      parameters:
        - name: global_id
          in: path
          type: string
          required: True
      responses:
        200:
          description: Successfully read service from service list 
          schema:
            properties:
              id:
                type: string
                description: ID of inter-site service
              type:
                type: string
                description: Type of inter-site service
              name:
                type: string
                description: Name of the inter-site service
              params:
                type: string
                description: Additional params for the inter-site service
              resources:
                type: array
                items:
                  type: string
              interconnections:
                type: array
                items:
                  type: string

  /intersite-horizontal:
    post:
      operationId: horizontal_create_service
      tags:
        - services
      summary: Horizontal request to create an inter-site Service POST
      description: Horizontal request to create an inter-site Service 
      x-swagger-router-controller: controllers.intersite
      parameters:
        - name: service
          in: body
          description: data for inter-site creation
          required: True
          schema:
            type: object
            properties:
              global:
                type: string
                description: First created ID of the service
              type:
                type: string
                description: Type of inter-site service
              name:
                type: string
                description: Name of the inter-site service
              params:
                type: string
                description: Additional params for the inter-site service
              resources:
                type: array
                items:
                  type: string 
      responses:
        201:
          description: Successful create inter-site service operation

  /intersite-horizontal/{global_id}:
    delete:
      operationId: horizontal_delete_service
      tags:
        - services
      summary: Deletes an inter-site Service DELETE
      description: Deletes an inter-site service
      x-swagger-router-controller: controllers.intersite
      parameters:
        - name: global_id
          in: path
          type: string
          required: True
          description: Id of the service to delete
      responses:
        200:
          description: Successfully deleted a service from service list